removed conftest.py (#231)
* Update CI.yml (#228)

runs-on: self-hosted (workstation-mcbs)

* Update CI.yml (#229)

* Update CI.yml

* Update CI.yml

* Update CI.yml

* removed conftest.py

* renamed common.py to conftest.py for pytest to recognize it

* not importing Flow to moments.py

* not importing Flow to inc_quad_eq
only import TorchStencil to _flow

* removed default Transform imports to avoid cyclic import

* pep8 cosmetics

---------
