+++ tests/__init__.py
@@ -1,3 +1,3 @@
 """Top-level package for tests."""
 
-__all__ = ["common"]
+__all__ = ["conftest"]
