+++ tests/collision/test_force.py
@@ -1,4 +1,4 @@
-from tests.common import *
+from tests.conftest import *
 import matplotlib.pyplot as plt
