+++ tests/moments/test_moment_equilibrium_dellar.py
@@ -1,4 +1,4 @@
-from tests.common import *
+from tests.conftest import *
 
 
 def test_moment_equilibrium_dellar(fix_device, fix_dtype):
