+++ tests/reporter/test_vtk_reporter_mask.py
@@ -1,6 +1,6 @@
 import os
 
-from tests.common import *
+from tests.conftest import *
 
 
 def test_vtk_reporter_mask(tmpdir):
