+++ tests/unit/test_consistency.py
@@ -1,4 +1,4 @@
-from tests.common import *
+from tests.conftest import *
 from tests.unit import create_default_unit_conversion
